2015, day 24
Not ideal, since theoretically it is possible that the "remaining"
presents are not equal-weight divisible by 2 (or 3, respectively).
However, with my input there were so many possible combinations that
this was considered unlikely, and it was indeed not a problem. One
way around this would have been to add the new result only after
checking that the remaining presents can be equal-weight divided into
2 or 3 groups.